+++ src/main/java/artemis/gui/DialogBox.java
@@ -1,4 +1,4 @@
-package duke.gui;
+package artemis.gui;
 
 import java.io.IOException;
 import java.util.Collections;
