+++ src/main/java/artemis/storage/TaskList.java
@@ -1,7 +1,7 @@
-package duke.storage;
+package artemis.storage;
 
-import duke.Deadline;
-import duke.Task;
+import artemis.Deadline;
+import artemis.Task;
 
 import java.time.LocalDate;
 import java.time.Period;
